+++ .gitattributes
@@ -1,0 +1,1 @@
+2014Q1-capitalbikeshare-tripdata.csv filter=lfs diff=lfs merge=lfs -text
